requirements.txt
astor==0.8.1
async-generator==1.10
Bottleneck==1.3.2
certifi==2021.10.8
cycler==0.10.0
entrypoints==0.3
fonttools==4.25.0
gast==0.2.2
h5py==2.10.0
joblib==1.1.0
jupyter==1.0.0
mistune==0.8.4
mkl-fft==1.3.1
mkl-service==2.4.0
munkres==1.1.4
numpy==1.21.4
olefile==0.46
opencv-python==4.5.4.60
pandas==1.3.4
Pillow==8.4.0
protobuf==3.17.2
pytz==2021.3
scikit-learn==1.0.1
six==1.16.0
tensorboard
tensorflow
tensorflow-estimator
termcolor==1.1.0
terminado==0.9.4
threadpoolctl==3.0.0
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1

